changed paths, deleted routes
