+++ package.json
@@ -25,6 +25,7 @@
     "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-hook-form": "^7.29.0",
     "react-loader-spinner": "^5.1.4",
     "react-query": "^3.34.15",
     "react-redux": "^7.2.6",
