+++ src/App.tsx
@@ -1,5 +1,4 @@
 import { Provider } from 'react-redux';
-import { BrowserRouter } from "react-router-dom";
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider } from '@mui/material/styles';
 import { persistStore } from 'redux-persist';
@@ -16,12 +15,10 @@ const App = () => {
   return (
     <Provider store={store}>
       <PersistGate persistor={persistor}>
-        <BrowserRouter>
           <ThemeProvider theme={theme}>
             <CssBaseline />
             <Main />
           </ThemeProvider>,
-        </BrowserRouter>
       </PersistGate>
     </Provider>
   )
